requirements.txt
# Core Dependencies
psycopg2-binary==2.9.9
requests==2.31.0
websocket-client==1.7.0
pandas==2.1.4
numpy==1.26.2

# Web Framework for Status UI
flask==3.0.0
flask-cors==4.0.0

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# Rate Limiting & Scheduling
schedule==1.2.1
ratelimit==2.2.1

# Logging
colorlog==6.8.0

# Proxy Support
PySocks==1.7.1
requests[socks]==2.31.0

# Data Validation
python-dotenv==1.0.0
pydantic==2.5.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
